+++ src/parameters/parameter.hpp
@@ -86,7 +86,7 @@ class effect_parameter {
         }
 
         virtual void try_gs_set_val() {
-            try_gs_effect_set_val(shader_param, data, data_size);
+            try_gs_effect_set_val(name.c_str(), shader_param, data, data_size);
         }
 
         /**