+++ src/shadertastic.hpp
@@ -20,7 +20,7 @@ typedef std::map<std::string, shadertastic_effect_t> shadertastic_effects_map_t;
 
 struct shadertastic_settings_t {
     // Path where the user can add their own effects
-    std::string *effects_path = NULL;
+    std::string *effects_path = nullptr;
     bool dev_mode_enabled = false;
 };
 #define SETTING_EFFECTS_PATH "effects_path"
@@ -31,7 +31,7 @@ shadertastic_settings_t shadertastic_settings;
 
 struct shadertastic_common {
     shadertastic_effects_map_t *effects;
-    shadertastic_effect_t *selected_effect = NULL;
+    shadertastic_effect_t *selected_effect = nullptr;
 };
 //----------------------------------------------------------------------------------------------------------------------
 
@@ -56,7 +56,7 @@ struct shadertastic_transition : public shadertastic_common {
         for (auto& [key, effect] : *this->effects) {
             effect.release();
         }
-        if (this->effects != NULL) {
+        if (this->effects != nullptr) {
             delete this->effects;
         }
     }